# Fix the critical performance issue of `crc32` function.