Fixes #
Fix the performance issue of `crc32` function.
The `crc32` function which implemented in Swift has a serious performance issue in debug build without compiler optimization. 
So it will affect the performance of compress and uncompress operation 

# Changes proposed in this PR
* I replaced the crc32 swift implementation with the C based implementation which included in the System Lib Zlib

### Performance Test 

#### Beta build [No compiler optimization]
Device : iPhone 11 Pro 
Text size 24.2 MB

|                     | compress  | uncompress |   
|---------------------|-----------|------------|
| Before Optimization | 13.067s   | 12.918s    |  
| After Optimization  | 0.348s    | 0.0949s    |
| Result  |  38x    | 143x    | 

#### Release build [optimize for speed [-O] ]
Device : iPhone 11 Pro 
Text size 24.2 MB

|                     | compress  | uncompress |   
|---------------------|-----------|------------|
| Before Optimization | 0.404s   | 0.165s    |  
| After Optimization  | 0.340s    | 0.0919 s    |
| Result  |  1.18x    | 1.795x    | 

#### Release build [optimize for size [-Osize] ]
Device : iPhone 11 Pro 
Text size 24.2 MB

|                     | compress  | uncompress |   
|---------------------|-----------|------------|
| Before Optimization | 0.409s   | 0.167s    |  
| After Optimization  | 0.332s    | 0.0908 s    |
| Result  |  1.23x    | 1.839x    | 
# Further info for the reviewer
Heaviest Stack trace from Instrument 

Before optimization
<img width="802" alt="Screen Shot 2020-09-25 at 17 19 00" src="https://user-images.githubusercontent.com/2288876/94246649-f58ab500-ff56-11ea-8423-fdfa176ceeda.png">

After optimization
<img width="842" alt="Screen Shot 2020-09-25 at 17 11 17" src="https://user-images.githubusercontent.com/2288876/94246661-f9b6d280-ff56-11ea-8124-a81d7452b6ab.png">

# Open Issues

- [ ] Can't build with Linux && Swift 4.0 environment