@noxo 2020-09-25
would be nice to get this PR in, i'm seeing 300x performance drops with crc32 on.