@noxo 2020-09-25
@andyqee Does this improve performance also in Release builds? I've noticed there is a ~300x performance drop in Debug builds, but then it Release mode haven't notice big difference (yet)? #135 #137