@weichsel 2020-09-25
Thank you for the PR and the benchmarks @andyqee .

As @noxo mentioned, the abysmal CRC32 performance is a symptom of Debug builds and should be reasonable when building with optimizations. 

On Darwin platforms, ZIP Foundation currently is self-contained and intentionally does not rely on zlib (trading performance during debugging for less dependencies).  
I'd prefer to keep it that way and instead of importing zlib, add a note to the README that explains the Debug vs. Release difference. It's also possible the improve the debugging experience by using a combination of `#if DEBUG` & `skipCRC32`.