@andyqee 2020-09-28
@noxo 
Thanks for you comments.
I tested with release build as you said the performance difference is not that big.

#### Release build [optimize for speed [-O] ]
Device : iPhone 11 Pro 
Text size 24.2 MB

|                     | compress  | uncompress |   
|---------------------|-----------|------------|
| Before Optimization | 0.404s   | 0.165s    |  
| After Optimization  | 0.340s    | 0.0919 s    |
| Result  |  1.18x    | 1.795x    | 

#### Release build [optimize for size [-Osize] ]
Device : iPhone 11 Pro 
Text size 24.2 MB

|                     | compress  | uncompress |   
|---------------------|-----------|------------|
| Before Optimization | 0.409s   | 0.167s    |  
| After Optimization  | 0.332s    | 0.0908 s    |
| Result  |  1.23x    | 1.839x    |