@weichsel 2020-09-28
@andyqee Thank you for re-doing your performance comparison with a release build!  
While the zlib based approach is slightly faster, I think this doesn't justify the additional dependency. 

I will close this PR and update the README to make people more aware of the huge performance discrepancy between Debug and Release builds.