@witekbobrowski 2021-01-21
I would also love to see this being mentioned in the README. We are using ZIPFoundation to zip user's data (including photos) for export and we did notice **huge** compress times in `DEBUG`. I am happy I finally landed here to learn about the real reason behind slow compression. Compressing in `RELEASE` is blazing fast.